REF: organize base class Index tests (#31864)
