+++ pandas/tests/indexes/test_common.py
@@ -158,13 +158,6 @@ class TestCommon:
         assert indices.name == name
         assert indices.names == [name]
 
-    def test_hash_error(self, indices):
-        index = indices
-        with pytest.raises(
-            TypeError, match=f"unhashable type: '{type(index).__name__}'"
-        ):
-            hash(indices)
-
     def test_copy_and_deepcopy(self, indices):
         from copy import copy, deepcopy
 
@@ -246,11 +239,6 @@ class TestCommon:
                 result = i._get_unique_index(dropna=dropna)
                 tm.assert_index_equal(result, expected)
 
-    def test_sort(self, indices):
-        msg = "cannot sort an Index object in-place, use sort_values instead"
-        with pytest.raises(TypeError, match=msg):
-            indices.sort()
-
     def test_mutability(self, indices):
         if not len(indices):
             pytest.skip("Skip check for empty Index")
@@ -260,9 +248,6 @@ class TestCommon:
 
     def test_view(self, indices):
         assert indices.view().name == indices.name
-
-    def test_compat(self, indices):
-        assert indices.tolist() == list(indices)
 
     def test_searchsorted_monotonic(self, indices):
         # GH17271
